refactor(chunkserver): Give own files to ChunkserverEntry
ChunkserverEntry now has its own .h and .cc files, which cleans up the
NetworkWorkerThread related files.

Misc changes in network_worker_thread.h and .cc:
- Remove unneeded includes.
- Use constexpr instead of macros.
- Wrap long lines.
